requirements.txt
numpy==1.18.4
PyQt5==5.15.0
PyQtChart==5.15.0
pigpio==1.46
PyYAML==5.3.1
scikit-learn==0.23.1
